chore: move `notiffake` to `testutil`
