+++ coderd/coderdtest/coderdtest.go
@@ -65,7 +65,6 @@ import (
 	"github.com/coder/coder/v2/coderd/gitsshkey"
 	"github.com/coder/coder/v2/coderd/httpmw"
 	"github.com/coder/coder/v2/coderd/notifications"
-	"github.com/coder/coder/v2/coderd/notifications/notiffake"
 	"github.com/coder/coder/v2/coderd/rbac"
 	"github.com/coder/coder/v2/coderd/schedule"
 	"github.com/coder/coder/v2/coderd/telemetry"
@@ -243,7 +242,7 @@ func NewOptions(t testing.TB, options *Options) (func(http.Handler), context.Can
 	}
 
 	if options.NotificationsEnqueuer == nil {
-		options.NotificationsEnqueuer = new(notiffake.FakeNotificationEnqueuer)
+		options.NotificationsEnqueuer = new(testutil.FakeNotificationEnqueuer)
 	}
 
 	accessControlStore := &atomic.Pointer[dbauthz.AccessControlStore]{}
